+++ src/components/cart/CartAmountGroup.jsx
@@ -1,55 +1,11 @@
-import styled from '@emotion/styled';
+import { currencyFomater, getSum, getPrices } from '../../util/commonUtils';
+import {
+  AmountBoxGruop,
+  AmountBox,
+  OrderButton,
+  Notice,
+} from '../../styles/CartPageStyle';
 
-import { currencyFomater } from '../../util/commonUtils';
-
-const AmountBoxGruop = styled.div({
-  width: '700px',
-  display: 'flex',
-  direction: 'row',
-  justifyContent: 'space-around',
-  marginTop: '40px',
-  // backgroundColor: 'palegreen',
-});
-const AmountBox = styled.div({
-  maxWidth: '100px',
-  alignItems: 'center',
-  display: 'flex',
-  flexDirection: 'column',
-  // backgroundColor: 'orange',
-  '& p': {
-    fontsize: '2em',
-  },
-  '& span': {
-    fontWeight: 'bolder',
-    fontsize: '1em',
-  },
-});
-const Button = styled.button({
-  backgroundColor: 'tomato',
-  color: 'white',
-  fontWeight: 'bold',
-  marginTop: '60px',
-  padding: '0.5rem 1rem',
-  fontSize: '1rem',
-  textAlign: 'center',
-  textDecoration: 'none',
-  display: 'inline-block',
-  width: '50%',
-  border: 'none',
-  borderRadius: '4px',
-});
-
-const Notice = styled.p({
-  color: '#C14',
-});
-
-function getSum(numbers) {
-  const sum = numbers.reduce((x, y) => x + y, 0);
-  return sum;
-}
-function getPrices(mCart) {
-  return mCart.map((item) => item.realPrice * item.itemAmount);
-}
 export default function CartAmountGroup({ cart }) {
   const sumPrices = getSum(getPrices(cart));
   const deliveryFee = sumPrices > 30000 ? 0 : 3000;
@@ -70,7 +26,7 @@ export default function CartAmountGroup({ cart }) {
           <span>{currencyFomater({ number: sumPrices + deliveryFee })}</span>
         </AmountBox>
       </AmountBoxGruop>
-      <Button type="button"> 주문하기</Button>
+      <OrderButton type="button"> 주문하기</OrderButton>
       <Notice> 30,000원 이하의 주문에는 배송비 3,000원이 추가됩니다.</Notice>
     </div>
   );
