+++ src/components/cart/CartItemRow.jsx
@@ -1,43 +1,12 @@
-import styled from '@emotion/styled';
-
 import { currencyFomater } from '../../util/commonUtils';
-
-const Item = styled.tr({
-  // backgroundColor: 'ivory',
-  height: '100px',
-  marginTop: '10px',
-  display: 'flow',
-  '& div': {
-    // backgroundColor: 'red',
-    maxWidth: '500px',
-    display: 'flex',
-    flexDirection: 'row',
-    justifyContent: 'space-around',
-    alignItems: 'center',
-    '& img': {
-      margin: '10px',
-      width: '70px',
-    },
-    '& p': {
-      width: '300px',
-      whiteSpace: 'nowrap',
-      overflow: 'hidden',
-      textOverflow: 'ellipsis',
-      display: 'block',
-      color: '#333',
-      textDecoration: 'none',
-      textAlign: 'center',
-      fontWeight: 'bolder',
-    },
-  },
-});
+import { ItemRow } from '../../styles/CartPageStyle';
 
 export default function CartItemRow({ item, onChangeCheckBox }) {
   const {
     id, name, img, itemAmount, realPrice, checked,
   } = item;
   return (
-    <Item>
+    <ItemRow>
       <td>
         <label htmlFor={`cartItem${id}`}>
           <input
@@ -64,6 +33,6 @@ export default function CartItemRow({ item, onChangeCheckBox }) {
         개
       </td>
       <td>{currencyFomater({ number: realPrice })}</td>
-    </Item>
+    </ItemRow>
   );
 }
